Extract module to file
