+++ crates/phactory/src/contracts/mod.rs
@@ -30,239 +30,4 @@ fn account_id_from_hex(s: &str) -> Result<AccountId> {
 }
 
 pub use support::*;
-mod support {
-    use phala_crypto::ecdh::EcdhPublicKey;
-    use phala_mq::traits::MessageChannel;
-    use runtime::BlockNumber;
-
-    use super::pink::group::GroupKeeper;
-    use super::*;
-    use crate::types::BlockInfo;
-
-    pub struct ExecuteEnv<'a> {
-        pub block: &'a BlockInfo<'a>,
-        pub contract_groups: &'a mut GroupKeeper,
-    }
-
-    pub struct NativeContext<'a> {
-        pub block: &'a BlockInfo<'a>,
-        pub mq: &'a SignedMessageChannel,
-        pub secret_mq: SecretMessageChannel<'a, SignedMessageChannel>,
-        pub contract_groups: &'a mut GroupKeeper,
-    }
-
-    pub struct QueryContext<'a> {
-        pub block_number: BlockNumber,
-        pub now_ms: u64,
-        pub contract_groups: &'a mut GroupKeeper,
-    }
-
-    impl NativeContext<'_> {
-        pub fn mq(&self) -> &SignedMessageChannel {
-            self.mq
-        }
-    }
-
-    pub trait Contract {
-        fn id(&self) -> ContractId;
-        fn handle_query(
-            &mut self,
-            origin: Option<&chain::AccountId>,
-            req: OpaqueQuery,
-            context: &mut QueryContext,
-        ) -> Result<OpaqueReply, OpaqueError>;
-        fn group_id(&self) -> Option<phala_mq::ContractGroupId>;
-        fn process_next_message(&mut self, env: &mut ExecuteEnv) -> Option<TransactionResult>;
-        fn on_block_end(&mut self, env: &mut ExecuteEnv) -> TransactionResult;
-        fn push_message(&self, payload: Vec<u8>, topic: Vec<u8>);
-        fn push_osp_message(
-            &self,
-            payload: Vec<u8>,
-            topic: Vec<u8>,
-            remote_pubkey: Option<&EcdhPublicKey>,
-        );
-        fn set_on_block_end_selector(&mut self, selector: u32);
-    }
-
-    pub trait NativeContract {
-        type Cmd: Decode + Debug;
-        type QReq: Decode + Debug;
-        type QResp: Encode + Debug;
-
-        fn id(&self) -> ContractId;
-        fn group_id(&self) -> Option<phala_mq::ContractGroupId> {
-            None
-        }
-        fn handle_command(
-            &mut self,
-            _origin: MessageOrigin,
-            _cmd: Self::Cmd,
-            _context: &mut NativeContext,
-        ) -> TransactionResult {
-            Ok(Default::default())
-        }
-        fn handle_query(
-            &mut self,
-            origin: Option<&chain::AccountId>,
-            req: Self::QReq,
-            context: &mut QueryContext,
-        ) -> Self::QResp;
-        fn on_block_end(&mut self, _context: &mut NativeContext) -> TransactionResult {
-            Ok(Default::default())
-        }
-        fn set_on_block_end_selector(&mut self, _selector: u32) {}
-    }
-
-    pub struct NativeCompatContract<Con, Cmd, CmdWrp, CmdPlr, QReq, QResp>
-    where
-        Cmd: Decode + Debug,
-        CmdWrp: Decode + Debug,
-        CmdPlr: Peeler<Wrp = CmdWrp, Msg = Cmd>,
-        QReq: Decode + Debug,
-        QResp: Encode + Debug,
-        Con: NativeContract<Cmd = Cmd, QReq = QReq, QResp = QResp>,
-    {
-        contract: Con,
-        send_mq: SignedMessageChannel,
-        cmd_rcv_mq: PeelingReceiver<Cmd, CmdWrp, CmdPlr>,
-        ecdh_key: KeyPair,
-    }
-
-    impl<Con, Cmd, CmdWrp, CmdPlr, QReq, QResp>
-        NativeCompatContract<Con, Cmd, CmdWrp, CmdPlr, QReq, QResp>
-    where
-        Cmd: Decode + Debug,
-        CmdWrp: Decode + Debug,
-        CmdPlr: Peeler<Wrp = CmdWrp, Msg = Cmd>,
-        QReq: Decode + Debug,
-        QResp: Encode + Debug,
-        Con: NativeContract<Cmd = Cmd, QReq = QReq, QResp = QResp>,
-        Cmd: Decode + Debug,
-    {
-        pub fn new(
-            contract: Con,
-            send_mq: SignedMessageChannel,
-            cmd_rcv_mq: PeelingReceiver<Cmd, CmdWrp, CmdPlr>,
-            ecdh_key: KeyPair,
-        ) -> Self {
-            NativeCompatContract {
-                contract,
-                send_mq,
-                cmd_rcv_mq,
-                ecdh_key,
-            }
-        }
-    }
-
-    impl<Con, Cmd, CmdWrp, CmdPlr, QReq, QResp> Contract
-        for NativeCompatContract<Con, Cmd, CmdWrp, CmdPlr, QReq, QResp>
-    where
-        Cmd: Decode + Debug,
-        CmdWrp: Decode + Debug,
-        CmdPlr: Peeler<Wrp = CmdWrp, Msg = Cmd>,
-        QReq: Decode + Debug,
-        QResp: Encode + Debug,
-        Con: NativeContract<Cmd = Cmd, QReq = QReq, QResp = QResp> + Send,
-        Cmd: Decode + Debug,
-    {
-        fn id(&self) -> ContractId {
-            self.contract.id()
-        }
-
-        fn group_id(&self) -> Option<phala_mq::ContractGroupId> {
-            self.contract.group_id()
-        }
-
-        fn handle_query(
-            &mut self,
-            origin: Option<&runtime::AccountId>,
-            req: OpaqueQuery,
-            context: &mut QueryContext,
-        ) -> Result<OpaqueReply, OpaqueError> {
-            debug!(target: "contract", "Contract {:?} handling query", self.id());
-            let response = self
-                .contract
-                .handle_query(origin, deopaque_query(req)?, context);
-            Ok(response.encode())
-        }
-
-        fn process_next_message(&mut self, env: &mut ExecuteEnv) -> Option<TransactionResult> {
-            let secret_mq = SecretMessageChannel::new(&self.ecdh_key, &self.send_mq);
-            let mut context = NativeContext {
-                block: env.block,
-                mq: &self.send_mq,
-                secret_mq,
-                contract_groups: &mut env.contract_groups,
-            };
-
-            phala_mq::select! {
-                next_cmd = self.cmd_rcv_mq => match next_cmd {
-                    Ok((_, cmd, origin)) => {
-                        info!(target: "contract", "Contract {:?} handling command", self.id());
-                        self.contract.handle_command(origin, cmd, &mut context)
-                    }
-                    Err(_e) => {
-                        Err(TransactionError::ChannelError)
-                    }
-                },
-            }
-        }
-
-        fn on_block_end(&mut self, env: &mut ExecuteEnv) -> TransactionResult {
-            let secret_mq = SecretMessageChannel::new(&self.ecdh_key, &self.send_mq);
-            let mut context = NativeContext {
-                block: env.block,
-                mq: &self.send_mq,
-                secret_mq,
-                contract_groups: &mut env.contract_groups,
-            };
-            self.contract.on_block_end(&mut context)
-        }
-
-        fn set_on_block_end_selector(&mut self, selector: u32) {
-            self.contract.set_on_block_end_selector(selector)
-        }
-
-        fn push_message(&self, payload: Vec<u8>, topic: Vec<u8>) {
-            self.send_mq.push_data(payload, topic)
-        }
-
-        fn push_osp_message(
-            &self,
-            payload: Vec<u8>,
-            topic: Vec<u8>,
-            remote_pubkey: Option<&EcdhPublicKey>,
-        ) {
-            let secret_mq = SecretMessageChannel::new(&self.ecdh_key, &self.send_mq);
-            secret_mq
-                .bind_remote_key(remote_pubkey)
-                .push_data(payload, topic)
-        }
-    }
-
-    pub use keeper::*;
-    mod keeper {
-        use phala_mq::ContractId;
-        use std::collections::BTreeMap;
-        use std::ops::{Deref, DerefMut};
-
-        type ContractMap = BTreeMap<ContractId, Box<dyn super::Contract + Send>>;
-
-        #[derive(Default)]
-        pub struct ContractsKeeper(ContractMap);
-
-        impl Deref for ContractsKeeper {
-            type Target = ContractMap;
-
-            fn deref(&self) -> &Self::Target {
-                &self.0
-            }
-        }
-
-        impl DerefMut for ContractsKeeper {
-            fn deref_mut(&mut self) -> &mut Self::Target {
-                &mut self.0
-            }
-        }
-    }
-}
+mod support;
